Portions Estimation/requirements.txt
# Project Dependencies
google-cloud-storage==2.18.2
pandas==2.2.3
opencv-python==4.10.0.84
numpy==2.1.3
tqdm==4.67.1